Inline test_infra.h
