+++ tests/test_skipdata.c
@@ -6,8 +6,29 @@
 #include "../myinttypes.h"
 
 #include <capstone.h>
-#include "test_infra.h"
 
+struct platform {
+	cs_arch arch;
+	cs_mode mode;
+	unsigned char *code;
+	size_t size;
+	char *comment;
+	cs_opt_type opt_type;
+	cs_opt_value opt_value;
+	cs_opt_type opt_skipdata;
+	size_t skipdata;
+};
+
+static void print_string_hex(unsigned char *str, size_t len)
+{
+	unsigned char *c;
+
+	printf("Code: ");
+	for (c = str; c < str + len; c++) {
+		printf("0x%02x ", *c & 0xff);
+	}
+	printf("\n");
+}
 
 #ifdef CAPSTONE_ARM_SUPPORT
 static size_t mycallback(const uint8_t *buffer, size_t buffer_size, size_t offset, void *p)
